+++ com.spoticket.ticket/src/main/java/com/spoticket/ticket/global/exception/GlobalExceptionHandler.java
@@ -1,5 +1,6 @@
 package com.spoticket.ticket.global.exception;
 
+import com.spoticket.ticket.global.util.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
