+++ com.spoticket.ticket/src/main/java/com/spoticket/ticket/presentation/TicketController.java
@@ -7,10 +7,10 @@ import com.spoticket.ticket.application.dtos.response.TicketInfoResponse;
 import com.spoticket.ticket.application.dtos.response.TicketResponse;
 import com.spoticket.ticket.application.service.TicketService;
 import com.spoticket.ticket.domain.entity.TicketStatus;
-import com.spoticket.ticket.global.exception.ApiResponse;
 import com.spoticket.ticket.global.exception.BusinessException;
 import com.spoticket.ticket.global.exception.ErrorCode;
-import com.spoticket.ticket.global.exception.PagedApiResponse;
+import com.spoticket.ticket.global.util.ApiResponse;
+import com.spoticket.ticket.global.util.PagedApiResponse;
 import jakarta.validation.Valid;
 import java.util.List;
 import java.util.UUID;
